OS11SPRYKE-1067_Edo - Fix Travis setup (#2324)
* Travis fix errors

* Travis fix errors

* Change travis.yml

* spryker files fixes

* adding to git

* travis change

* travis change

* fixes

* fixes

* fixes

* fixes

* fixes

* travis phpstan memorylimits

* added rabbitmq-server

* travis fix

* travis fix

* travis fix

* travis fix

* config files change

* config files change

* [OS11SPRYKE-1067] - commit

* Code sniffer fix.

* Changes in the Travis file. Change in the Index file.

* Reverting PHP version

* Fixing queue problems

* Fixing DB version problem

* [OS11SPRYKE-1067] - Neon Commit.

* [OS11SPRYKE-1067] - Neon Commit.

* [OS11SPRYKE-1067] - Neon Commit.

* [OS11SPRYKE-1067] - Neon Commit 5.

* [OS11SPRYKE-1067] - Neon Commit 6.

* [OS11SPRYKE-2071] - Bugfix.

* [OS11SPRYKE-1067] - Bugfix.

* [OS11SPRYKE-1067] - commit #2

* [OS11SPRYKE-1067] - commit #3

* [OS11SPRYKE-1067] - commit #4

* [OS11SPRYKE-1067] - Commit #5

* [OS11SPRYKE-1067] - Commit #6

* [OS11SPRYKE-1067] - Commit #7

* [OS11SPRYKE-1067] - Commit #8

* [OS11SPRYKE-1067] - Commit #9

* [OS11SPRYKE-1067] - Commit #10

* [OS11SPRYKE-1067] - Commit #13

* [OS11SPRYKE-1067] - Commit #14

* [OS11SPRYKE-1067] - Commit #14

* [OS11SPRYKE-1067] - Commit #15

* Testing valiue

* Tried to remove Application test from SpyTests

* [OS11SPRYKE-1067] - Commit 18

* [OS11SPRYKE-1067] - Commit 19

* [OS11SPRYKE-1067] - Commit 20

* [OS11SPRYKE-1067] - Commit 21

* [OS11SPRYKE-1067] - Commit 22

* [OS11SPRYKE-1067] - Commit 23

* [OS11SPRYKE-1067] - Commit 25

* [OS11SPRYKE-1067] - Commit 26

* [OS11SPRYKE-1067] - Commit 27

* [OS11SPRYKE-1067] - Commit 28

* [OS11SPRYKE-1067] - Commit 29

* [OS11SPRYKE-1067] - Commit 30

* [OS11SPRYKE-1067] - Commit 31

* [OS11SPRYKE-1067] - Commit 32

* [OS11SPRYKE-1067] - phpstan && code sniff fix

* [OS11SPRYKE-1067] - phpstan && code sniff fix 2

* [OS11SPRYKE-1067] - phpstan && code sniff fix 3

* [OS11SPRYKE-1067] - phpstan && code sniff fix 5

+++ src/Pyz/Yves/CustomerPage/Controller/ProfileController.php
@@ -132,9 +132,9 @@ class ProfileController extends SprykerShopProfileController
     }
 
     /**
-     * @param int|string|null $iso2Code
+     * @param string|null $iso2Code
      *
-     * @return string
+     * @return int
      */
     public function getIdCountryFromISO2CodeFromConfig(?string $iso2Code): int
     {
@@ -143,7 +143,7 @@ class ProfileController extends SprykerShopProfileController
             return 60;
         } else {
             if ($countries != null) {
-                return array_search($iso2Code, $countries);
+                return (int)array_search($iso2Code, $countries);
             } else {
                 return 60;
             }
